+++ resources/views/Position/newIndex.blade.php
@@ -26,6 +26,7 @@
                     <thead class="bg-primary">
                         <tr>
                             <th>Position</th>
+                            <th>Status</th>
                             <th class="noSortAction">Action</th>
                         </tr>
                     </thead>  
@@ -33,6 +34,22 @@
                         @foreach($positions as $positions)
                             <tr>
                                 <td>{{$positions->strPositionName}}</td>
+                                <td>
+                                    @if($positions->isActive == 1)
+                                        <div class="positionCheckbox" data-id="{{$positions->intPositionID}}">
+                                            <label>
+                                                <input type="checkbox" checked data-size="small" data-width="" data-toggle="toggle"data-onstyle="success"data-offstyle="danger" data-on="Active" data-off="Inactive" data-id="{{$positions->intPositionID}}">
+                                            </label>
+                                        </div>
+                                    @elseif($positions->isActive == 0)
+                                        <div class="positionCheckbox" data-id="{{$positions->intPositionID}}">
+                                            <label> 
+                                                <input type="checkbox" data-size="small" data-width="" data-toggle="toggle"data-onstyle="success"data-offstyle="danger" data-on="Active" data-off="Inactive" data-id="{{$positions->intPositionID}}">
+                                            </label>
+                                        </div>
+                                    @endif
+                                    
+                                </td>
                                 <td>
                                     <div class="ml-1 mr-1">
                                         <button onclick="getPosition({{$positions->intPositionID}})"class="btn btn-sm btn-secondary" data-toggle="tooltip" title="Edit" role="button">
@@ -52,6 +69,7 @@
                     <thead class="bg-primary">
                         <tr>
                             <th>Position</th>
+                            <th></th>
                             <th class="noSortAction">Action</th>
                         </tr>
                     </thead>